requirements_session_checker.txt
# متطلبات SessionChecker.py
python-telegram-bot==13.15
telethon>=1.24.0
pyrogram>=2.0.0
requests>=2.28.0